+++ Disk/Calculations/Impl/Calculator2D.cs
@@ -4,19 +4,19 @@ namespace Disk.Calculations.Impl
 {
     static class Calculator2D
     {
-        public static Point2D<double> MathExp(IEnumerable<Point2D<double>> dataset) => new
+        public static Point2D<float> MathExp(IEnumerable<Point2D<float>> dataset) => new
             (
                 Calculator.MathExp(dataset.Select(p => p.X)),
                 Calculator.MathExp(dataset.Select(p => p.Y))
             );
 
-        public static Point2D<double> StandartDeviation(IEnumerable<Point2D<double>> dataset) => new
+        public static Point2D<float> StandartDeviation(IEnumerable<Point2D<float>> dataset) => new
             (
                 Calculator.StandartDeviation(dataset.Select(p => p.X)),
                 Calculator.StandartDeviation(dataset.Select(p => p.Y))
             );
 
-        public static Point2D<double> Dispersion(IEnumerable<Point2D<double>> dataset) => new
+        public static Point2D<float> Dispersion(IEnumerable<Point2D<float>> dataset) => new
             (
                 Calculator.Dispersion(dataset.Select(p => p.X)),
                 Calculator.Dispersion(dataset.Select(p => p.Y))
